requirements.txt
django==5.0.8
djangorestframework==3.15.2
psycopg[binary]==3.2.1
django-environ==0.11.2
dj-rest-auth[with_social]==6.0.0
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.14.1
drf-yasg==1.21.7
setuptools==74.0.0
django-filter==24.3
